Move builtins to their own header, specializations to their respective declarations
